hack/requirements.txt
Django==5.2.6
Pillow==10.4.0
qrcode==8.2
numpy==2.2.6
pandas==2.3.2
opencv-python==4.8.1.78
